feat: improve testing
